fix: move Swipeable type definition
