+++ package.json
@@ -43,6 +43,7 @@
     "PlatformConstants.js",
     "PlatformConstants.web.js",
     "react-native-gesture-handler.d.ts",
+    "Swipeable.d.ts",
     "README.md",
     "RNGestureHandlerModule.js",
     "RNGestureHandlerModule.web.js",
